backend/pyproject.toml
[project]
name = "falcon"
version = "0.1.0"
description = "Open-source AI wiki generator for GitHub repositories"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "aiosqlite>=0.20.0",
    "httpx>=0.27.0",
    "sse-starlette>=2.0",
    "python-frontmatter>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
]
daytona = [
    "daytona-sdk>=0.1.0",
]

[build-system]
requires = ["setuptools>=70.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
